做moses介面的試驗而且實作了
